% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\alias{accuracy}
\title{Calculate confusion matrix and accuracy for Net Promoter Category
predicitons}
\usage{
confusion_matrix(m, test_data)

accuracy(m, test_data)
}
\arguments{
\item{m}{Model output from \code{model_survey}}

\item{test_data}{A data frame with test data in the same format as the input
data for the model object}
}
\description{
Calculate confusion matrix and accuracy for Net Promoter Category
predicitons
}
\examples{
library(dplyr)
library(NPS)
library(caret)

data(survey_sim)

survey_sim <- survey_sim \%>\%
  mutate(NPC = factor(npc(recommend), ordered = TRUE)) \%>\%
  select(-recommend)

index <- createDataPartition(survey_sim$NPC, p = .8, list = FALSE, times = 1)
train <- survey_sim[index, ]
test <- survey_sim[-index, ]

m <- model_survey(train)

confusion_matrix(m, test)
accuracy(m, test)
}
