% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scenarios.R
\name{make_scenarios}
\alias{make_scenarios}
\title{Make table of scenarios}
\usage{
make_scenarios(m, redistribution = 0.5)
}
\arguments{
\item{m}{Model output from \code{model_survey()}}

\item{redistribution}{Proportion of highest score distributed to lower scores}
}
\description{
Make table of scenarios
}
\examples{
library(NPS)
library(NPSdrivers)
library(dplyr)

data(survey_sim)

survey_sim <- survey_sim \%>\%
  mutate(NPC = factor(npc(recommend), ordered = TRUE)) \%>\%
  select(-recommend)

m <- model_survey(survey_sim)
scenarios <- make_scenarios(m)
}
