% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_survey.R
\name{model_survey}
\alias{model_survey}
\title{Fit ordered logistic regression model to survey data}
\usage{
model_survey(data, verbose = FALSE)
}
\arguments{
\item{data}{A data frame of survey results with an "NPC" column (for Net Promoter
Catagory). See example.}

\item{verbose}{Logical indicating if you want to see output from \code{stepAIC()}}
}
\description{
Fit ordered logistic regression model to survey data
}
\examples{
library(NPS)
library(dplyr)

data(survey_sim)

survey_sim <- survey_sim \%>\%
  mutate(NPC = factor(npc(recommend), ordered = TRUE)) \%>\%
  select(-recommend)

m <- model_survey(survey_sim)
}
