% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.NPSdriverScenarios}
\alias{plot.NPSdriverScenarios}
\title{Plot NPS Driver Scenarios}
\usage{
\method{plot}{NPSdriverScenarios}(x, NPS, ...)
}
\arguments{
\item{x}{Output from \code{run_scenarios()}}

\item{NPS}{The current NPS in the data}

\item{...}{Parameters passed on to \code{aes()}}
}
\description{
Plot NPS Driver Scenarios
}
\examples{
library(NPS)
library(NPSdrivers)
library(dplyr)

data(survey_sim)

NPS <- NPS::nps(survey_sim$recommend)
survey_sim <- survey_sim \%>\%
  mutate(NPC = factor(npc(recommend), ordered = TRUE)) \%>\%
  select(-recommend)

m <- model_survey(survey_sim)
scenario_results <- run_scenarios(m)
plot(scenario_results, NPS)
}
