% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_sim.R
\docType{data}
\name{survey_sim}
\alias{survey_sim}
\title{Simulated survey questions.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{recommend}{would you recommend, scale 0 to 10}
  \item{corr1}{correlated categorical variable, scale 1 to 7}
  \item{corr2}{correlated categorical variable, scale 1 to 7}
  \item{uncorr}{uncorrelated categorical variable, scale 1 to 7}
}}
\usage{
survey_sim
}
\description{
A dataset containing simulated Likert scale questions that are correlated
and uncorrelated.
}
\keyword{datasets}
