% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.NPSdrivers}
\alias{tidy.NPSdrivers}
\alias{augment.NPSdrivers}
\alias{glance.NPSdrivers}
\title{Tidy up the output from \code{model_survey()}}
\usage{
\method{tidy}{NPSdrivers}(m)

\method{augment}{NPSdrivers}(m)

\method{glance}{NPSdrivers}(m)
}
\arguments{
\item{m}{Model output from \code{model_survey}}
}
\description{
Tidy up the output from \code{model_survey()}
}
\examples{
library(NPS)
library(dplyr)
library(broom)

data(survey_sim)

survey_sim <- survey_sim \%>\%
  mutate(NPC = factor(npc(recommend), ordered = TRUE)) \%>\%
  select(-recommend)

m <- model_survey(survey_sim)

tidy(m)
head(augment(m))
glance(m)
}
