% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scenarios.R
\name{run_scenarios}
\alias{run_scenarios}
\title{Run different scenarios on the survey data based on the model output from
\code{model_survey()}}
\usage{
run_scenarios(model, scenarios = make_scenarios(model))
}
\arguments{
\item{model}{Model object from \code{model_survey()}}

\item{scenarios}{A data frame of scenarios
library(NPS)
library(NPSdrivers)
library(dplyr)

data(survey_sim)

survey_sim <- survey_sim %>%
  mutate(NPC = factor(npc(recommend), ordered = TRUE)) %>%
  select(-recommend)

m <- model_survey(survey_sim)
scenario_results <- run_scenarios(m)}
}
\description{
Run different scenarios on the survey data based on the model output from
\code{model_survey()}
}
